% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airExample.r
\name{airExample}
\alias{airExample}
\title{Obtain example data that comes with airUpThere}
\usage{
airExample(x)
}
\arguments{
\item{x}{Character: The name of a data set:
\itemize{
\item \link{parks}: Outline of two parks in the City of Saint Louis, Missouri, USA ("polygons" \code{SpatVector})
\item \link{stl}: Outline of the City of Saint Louis
\item \link{trees}: Randomly sampled locations of trees managed by the City of Saint Louis ("points", \code{SpatVector})
}}
}
\value{
A \code{SpatVector} or \code{SpatRaster}.
}
\description{
This function makes retrieving example data that comes with \strong{airUpThere} easy.
}
\examples{

stl <- airExample('stl')
stl
plot(stl)

}
