% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airUpdate.r
\name{airUpdate}
\alias{airUpdate}
\title{Update URLs and variable names used by airUpThere}
\usage{
airUpdate(verbose = TRUE)
}
\value{
Creates objects named \code{airUrls} and \code{airVars}. Invisibly returns \code{TRUE} or \code{FALSE}, depending on success.
}
\description{
The purpose of this function is to enable easy updates to new versions of climate data sources as they become available without having to update the entire package. The function downloads the latest version of URLs, file name patterns, and variable names used by \code{airUpThere}. It is not necessary to run this function before using \pkg{airUpThere}, but if you find that files will not download using a \code{xxDownload} function, please try running this function then trying again.
}
\examples{

airUpdate()

}
