% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airUrls.r
\docType{data}
\name{airUrls}
\alias{airUrls}
\title{Names, abbreviations, version numbers, and URLs for each climate data source}
\format{
An object of class \code{'data.frame'}.
}
\usage{
data(airUrls)
}
\description{
Names/abbreviations, URLs for raster files, URLs for timestamp files, and file pattern names for climate data sources. Please note that it is advisable to update \code{airUrls} using \code{airUpdate()} before running most \pkg{airUpThere} functions.
}
\examples{
data(airUrls)
airUrls
}
\seealso{
\code{\link[airUpThere]{updateAir}}
}
\keyword{variables}
