% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chInternalUtils.r
\name{chCheckEsm_internal}
\alias{chCheckEsm_internal}
\title{Check ESM for CHELSA}
\usage{
chCheckEsm_internal(cmip, esm)
}
\arguments{
\item{cmip}{CMIP (5 or 6)}

\item{esm}{Name of the ESM. Case-sensitive!}
}
\value{
\code{TRUE} (invisible) or an error will be called.
}
\description{
Check ESM for CHELSA
}
\keyword{internal}
