% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chConvertPeriod.r
\name{chConvertPeriod}
\alias{chConvertPeriod}
\title{Convert CHELSA "future" period shorthand to full range, or vice versa.}
\usage{
chConvertPeriod(cmip, period)
}
\arguments{
\item{cmip}{Number the Coupled Model Intercomparison Project (CMIP) future ratsers to download. Valid values are either \code{5} or \code{6}.}

\item{period}{Time period. Looks either like \code{2041} or \code{2041-2070}. Short/long versions should not be mixed.}
}
\value{
Character or integer.
}
\description{
Convert CHELSA "future" period shorthand (e.g., \code{2071}) to full range (e.g., \code{2071-2100}), or vice versa. The function will try to guess whether you want to go from short to long or vice versa.
}
\examples{

chConvertPeriod(5, 2041)
chConvertPeriod(5, c(2041, 2061))
chConvertPeriod(5, '2061-2080')

chConvertPeriod(6, 1981)
chConvertPeriod(6, c(1981, 2071))
chConvertPeriod(6, '2011-2040')
}
