% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chDownload.r
\name{chDownloadPaleoBio}
\alias{chDownloadPaleoBio}
\title{Download CHELSA climate rasters}
\usage{
chDownloadPaleoBio(
  saveTo,
  ver,
  bios,
  centuries,
  forceUpdate = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{saveTo}{Name of the base path to which to save the download. Sub-folders will be created within this folder.}

\item{ver}{Version number. Valid version numbers depend on whether paleo, historical, or future coverages are desired:
\itemize{
\item Paleo (TraCE2K): \code{1.0}
\item Historical: \code{XYZ}
\item Future under CMIP5: \code{1.2}
\item Future under CMIP6: \code{2.1}
}}

\item{bios}{The number of BIOCLIM variable(s) to get (any integers in \code{1:19}).}

\item{centuries}{Century or centuries from which to download the paleo-climate/elevation/ice rasters. This can be a value from -200 (the -200th century, or 21000 to 20001 ybp) to 20 (i.e., the 20th century, 1900-1999 CE).}

\item{forceUpdate}{If \code{FALSE} (default), then all existing rasters will be overwritten, regardless of whether or not the existing ones are up-to-date or not.}

\item{verbose}{If \code{TRUE} (default), display progress.}

\item{vars}{Name(s) of variable(s) to download. Valid values depend on the version of WorldClim and whether near-present day or future rasters are fetched. Different versions and time periods of WorldClim use different names for the same variable (e.g., "prec" versus "ppt" versus "pr" for precipitation). To reduce confusion, variable names have been standardized (in this package) to be the same across versions and times. Valid values are:
\itemize{
\item \code{elev}: elevation (function \code{chDownloadPaleoElev()})
\item \code{iceMask}: mask for ice cover (\code{1} for ice or \code{NA} for none) (function \code{chDownloadPaleoElev()})
\item \code{iceElev}: elevation of ice (plus land) (function \code{chDownloadPaleoElev()})
\item \code{snowCoverDays}: days of snow cover (function \code{chDownloadPaleoClim()})
\item \code{swe}: snow water equivalent (function \code{chDownloadPaleoClim()})
\item \code{tmin}: minimum temperature (function \code{chDownloadPaleoClim()})
\item \code{tmax}: maximum temperature (function \code{chDownloadPaleoClim()})
\item \code{ppt}: accumulated precipitation (function \code{chDownloadPaleoClim()})
}}

\item{cmip}{Number the Coupled Model Intercomparison Project (CMIP) future rasters to download. Valid values are either \code{5} or \code{6}.}

\item{esm}{Abbreviations of available earth system models, depending on whether CMIP5 or CMIP6 is desired. A list of valid options can be obtained using \code{\link[=chEsm]{chEsm()}}.}

\item{ghg}{Name of the emissions scenario of future rasters. Valid values depend on \code{cmip} (CMIP). For CMIP5 These are:
\itemize{
\item \code{2.6}: RCP2.6
\item \code{4.5}: RCP4.5
\item \code{6.0}: RCP6.0
\item \code{8.5}: RCP8.5
}

For CMIP6 these are:
\itemize{
\item \code{126}: SSP126
\item \code{370}: SSP370
\item \code{585}: SSP585
}}

\item{period}{Period from which to earth system model climate predictions (which includes some "historic" predictions). The first year of the respective time period is used. Depending on \code{cmip} (CMIP), different values are valid. A list of valid values can be see using \code{\link[=chPeriod]{chPeriod()}}}

\item{months}{The number of month(s) to get (any integers in \code{1:12}).}
}
\description{
These functions download \href{https://chelsa-climate.org/}{CHELSA} climate rasters, elevation, and permanent ice/snow cover and elevation thereof:
\itemize{
\item \code{chDownloadPaleoBio()}: Paleo/future climate BIOCLIM variables.
\item \code{chDownloadPaleoClim()}: Paleo climate "basic" variables (precipitation, temperature).
\item \code{chDownloadPaleoElev()}: Paleo elevation and ice cover.
\item \code{chDownloadFutureBio()}: Paleo/future climate BIOCLIM variables.
\item \code{chDownloadSeries()}: Monthly time series.
`
}
}
\examples{

\dontrun{
dl <- 'C:/ecology/!Scratch/tc'
chDownloadMonthly(dl, 'tmin', 1958)
chDownloadElev(dl, 'elev')

}
}
\references{
Karger, D. N., Nobis, M. P., Normand, S., Graham, C. H., & Zimmermann, N. E. (2021): CHELSA-TraCE21k v1. 0. Downscaled transient temperature and precipitation data since the last glacial maximum. \emph{Climate of the Past Discussions} 1-27. \doi{10.5194/cp-2021-30}.
}
