% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chEsm.r
\name{chEsm}
\alias{chEsm}
\title{Get names of CHELSA earth system models}
\usage{
chEsm(cmip)
}
\arguments{
\item{cmip}{Number the Coupled Model Intercomparison Project (CMIP) future ratsers to download. Valid values are either \code{5} or \code{6}.}
}
\value{
Character vector.
}
\description{
Return a vector of earth system model predictions available from CHELSA.
}
\examples{
chEsm(5)
chEsm(6)
}
