% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chInternalUtils.r
\name{chIsPeriodLong_internal}
\alias{chIsPeriodLong_internal}
\title{Are periods in short or long format?}
\usage{
chIsPeriodLong_internal(period)
}
\value{
\code{TRUE} (invisible) or an error will be called.
}
\description{
Are periods in short or long format?
}
\keyword{internal}
