% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chPeriod.r
\name{chPeriod}
\alias{chPeriod}
\title{Get time periods for CHELSA future projections}
\usage{
chPeriod(cmip)
}
\arguments{
\item{cmip}{Number the Coupled Model Intercomparison Project (CMIP) future ratsers to download. Valid values are either \code{5} or \code{6}.}
}
\value{
Character vector.
}
\description{
Return a data frame with time periods for CHELSA future projections.
}
\examples{
chPeriod(5)
chPeriod(6)
}
