% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDates.r
\name{compareDates}
\alias{compareDates}
\alias{\%>d\%}
\alias{\%>=d\%}
\alias{\%==d\%}
\alias{\%<d\%}
\alias{\%<=d\%}
\alias{\%!=d\%}
\title{Compare dates in YYYY, YYYY-MM, or YYYY-MM-DD formats}
\usage{
compareDates(d1, d2, op)

d1 \%>d\% d2

d1 \%>=d\% d2

d1 \%==d\% d2

d1 \%<d\% d2

d1 \%<=d\% d2

d1 \%!=d\% d2
}
\arguments{
\item{d1, d2}{Dates in YYYY, YYYY-MM, YYYY-MM-DD.}

\item{op}{Any of \code{'>'}, \code{'>='}, \code{'=='}, \code{'<'}, \code{'<='}, or \code{'!='}.}
}
\value{
Logical.
}
\description{
Compare dates in YYYY-MM format
}
\section{Functions}{
\itemize{
\item \code{d1 \%>d\% d2}: Compare dates of YYYY, YYYY-MM, or YYYY-MM-DD format using >

\item \code{d1 \%>=d\% d2}: Compare dates of YYYY, YYYY-MM, or YYYY-MM-DD format using >=

\item \code{d1 \%==d\% d2}: Compare dates of YYYY, YYYY-MM, or YYYY-MM-DD format using ==

\item \code{d1 \%<d\% d2}: Compare dates of YYYY, YYYY-MM, or YYYY-MM-DD format using <

\item \code{d1 \%<=d\% d2}: Compare dates of YYYY, YYYY-MM, or YYYY-MM-DD format using <=

\item \code{d1 \%!=d\% d2}: Compare dates of YYYY, YYYY-MM, or YYYY-MM-DD format using !=

}}
\examples{
compareDates('2021-04-27', '2021-05-27', '>')
compareDates('2021-05-27', '2021-04-27', '>')

compareDates('2021-05-27', '2021-05-27', '==')
compareDates('2021-04-27', '2021-05-27', '!=')

'2021-05-27' \%>d\% '2021-04-27'
}
