% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{fillPattern_internal}
\alias{fillPattern_internal}
\title{Fill in wildcards in file/URL names}
\usage{
fillPattern_internal(x)
}
\arguments{
\item{x}{Character (usually a URL or file name)}

\item{...}{Named arguments of values to fill in. Must be named any of: \code{ver}, \code{varFile}, \code{resFile}, \code{periodFile}, \code{ghgFile}, \code{esmFile}.}
}
\value{
Character.
}
\description{
This function replaces wildcards in generic file/URL names with the proper values. It first gathers any named wildcards (e.g., "varFile", "esmFile"), then for each one that occurs in a file/URL, name, replaces it with the value of the named object. The function basically translates generic names in \code{airUrls} to useful file names or URLs.
}
\keyword{internal}
