% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatYYYYMM.r
\name{formatYYYYMM}
\alias{formatYYYYMM}
\title{Get "YYYY-MO" (year-month) from a date.}
\usage{
formatYYYYMM(x)
}
\arguments{
\item{x}{Object(s) of class \code{Date}, or of class \code{character}.}
}
\value{
A character vector.
}
\description{
This function returns the "YYYY-MM" format of a date (i.e., dropping the "day").
}
\examples{
formatYYYYMM(c('2015-10-24', '2017-02-26', '2021-04'))
}
