% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCoords.r
\name{getCoords}
\alias{getCoords}
\title{Get coordinates}
\usage{
getCoords(x, index = NULL, longLat = NULL)
}
\arguments{
\item{x}{Object of class \code{data.frame}, \code{sf}, \code{SpatialPoints}, \code{SpatialPointsDataFrame}, or \code{SpatVector}.}

\item{index}{Either \code{NULL} (default) or an index of which coordinates to get. If \code{NULL}, all coordinates are obtained.}

\item{longLat}{Either \code{NULL} (default) or a two-element character vector with the names of the fields in a matrix or data frame with longitude and latitude (in that order).}
}
\description{
Get coordinates from an object of class \code{data.frame}, \code{matrix}, \code{SpatialPoints}, \code{SpatialPointsDataFrame}, or \code{SpatVector}.
}
