% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFileOrURL.r, R/utils.r
\name{getFileOrURL_internal}
\alias{getFileOrURL_internal}
\title{Get URL and file pattern name for a data source}
\usage{
getFileOrURL_internal(what, src, ver = NULL, period = NULL, elevation = FALSE)

getFileOrURL_internal(what, src, ver = NULL, period = NULL, elevation = FALSE)
}
\arguments{
\item{what}{Either \code{rastUrl}, \code{rastFilePattern}, \code{timestampFileName}, or \code{timestampFilePattern}.}

\item{src}{Name or abbreviation for climate data source}

\item{ver}{Version number, name, or \code{NULL}.}

\item{period}{Character or year indicating time period from which data is to be downloaded, or \code{NULL}.}

\item{elevation}{\code{TRUE} or \code{FALSE}.}
}
\value{
URL or file name pattern

URL or file name pattern
}
\description{
This function gets the appropriate URL and file name pattern for a given climate data source. Typically it is called within an \pkg{airUpThere} function.

This function gets the appropriate URL and file name pattern for a given climate data source. Typically it is called within an \pkg{airUpThere} function.
}
