% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYMD.r
\name{getYMD}
\alias{getYMD}
\title{Year, month, or day from a date}
\usage{
getYMD(d, get)
}
\arguments{
\item{d}{Date(s) in YYYY, YYYY-MM, or YYYY-MM-DD format.}

\item{get}{Any of \code{'year'}, \code{'month'}, or \code{'day'}.}
}
\value{
Integer
}
\description{
Get year, month, or day as an integer from a date without having to convert it to a \code{Date}.
}
\examples{
getYMD('2021-04-27', 'year')
getYMD('2021-04-27', 'month')
getYMD('2021-04-27', 'day')
}
