% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghgNames.r
\docType{data}
\name{ghgNames}
\alias{ghgNames}
\title{Names of future climate change scenarios (RCPs and SSPs)}
\format{
An object of class \code{'list'}.
}
\usage{
data(ghgNames)
}
\description{
Names of RCPs from CMIP5 and SSPs from CMIP6
}
\examples{
data(ghgNames)
ghgNames
}
\keyword{variables}
