% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthAsNum.r
\name{monthAsNum}
\alias{monthAsNum}
\title{Convert month name to number}
\usage{
monthAsNum(mo, names = TRUE)
}
\arguments{
\item{mo}{Character vector of month names.}

\item{names}{If \code{TRUE} (default), name the output vector by input names.}
}
\value{
Integer vector.
}
\description{
Convert a month name to a number.  Kinds of names supported include, for example, "October", "Oct", and "Oct.".  Unambiguous matching is used. For example, "J", "Ja", and "Janua" would all be matched to January so would return a 1 (first month).  "M" and "Ma" would not be matched since they could both stand for "March" or "May".
}
\examples{

mo <- c('jan', 'Jan', 'Jan.', 'Janu', 'january')
monthAsNum(mo)

mo <- c('j', 'f', 'm', 'a', 'm', 'j', 'j', 'a', 's', 'o', 'n', 'd')
monthAsNum(mo)
monthAsNum(mo, FALSE)

}
