% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthDiff.r
\name{monthDiff}
\alias{monthDiff}
\title{Difference in months between dates in YYYY-MM format}
\usage{
monthDiff(d1, d2)
}
\arguments{
\item{d1, d2}{Two dates (characters) expressed in YYYY-MM format.}
}
\value{
Integer indicating number of months difference between the two dates.
}
\description{
This function allows you to calculate the difference between two dates in YYYY-MM format.
}
\examples{

monthDiff('2020-10', '2020-05')
monthDiff('2020-10', '2019-10')
monthDiff('2020-10', '2017-12')

}
