% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nad83.r
\name{nad83}
\alias{nad83}
\title{Well-Known Text string for the NAD83 Datum}
\format{
\code{character}
}
\description{
This is the Well-Known Text (WKT) string for the North American Datum 1983 (NAD83).
}
\examples{

data(nad83)
nad83
cat(nad83)

}
\seealso{
\link{wgs84}
}
