% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parks.r
\name{parks}
\alias{parks}
\title{Outlines of two parks in Saint Louis, Missouri}
\format{
\code{SpatVector}
}
\source{
\href{https://www.stlouis-mo.gov/}{City of Saint Louis}
}
\description{
This file is a \code{SpatVector} outline of two parks in Saint Louis, Missouri. Forest Park is the largest municipal park in the US, and Tower Grove Park the largest Victorian park.
}
\examples{

parks <- airExample('parks')
parks
plot(parks)

}
\seealso{
\code{\link[=airExample]{airExample()}}
}
