% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prUtils.r
\name{prGetDates}
\alias{prGetDates}
\title{Get dates from a vector of a data frame-like object}
\usage{
prGetDates(x, date)
}
\arguments{
\item{x}{Data frame, SPDF, of object of class \code{SpatVector}}

\item{date}{Object of class{Date} or character vector with one or three elements}
}
\value{
Vector of \code{Date} objects
}
\description{
Get dates from a vector of a data frame-like object
}
\keyword{internal}
