% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prUtils.r
\name{prGetRes}
\alias{prGetRes}
\title{Convert PRISM resolution to name used in file names}
\usage{
prGetRes(res)
}
\arguments{
\item{res}{Either 30 (30 arcsec) or 800 (30 arcsec), or 1 (1 arcmin) or 4 (1 arcmin)}
}
\value{
Character
}
\description{
Convert PRISM resolution to name used in file names
}
\keyword{internal}
