% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prVars.r
\name{prVars}
\alias{prVars}
\title{Get Names of PRISM variables}
\usage{
prVars(ver, standard = TRUE, defs = FALSE)
}
\arguments{
\item{ver}{Version of PRISM. Either of \code{'public'} or \code{'subscription'}. Partial matching is used.}

\item{standard}{If \code{TRUE} (default), the "standard" names are returned. If \code{FALSE}, the file/URL names are returned.}

\item{defs}{If \code{TRUE}, return general definitions of each variable, too.}
}
\value{
A character vector.
}
\description{
Get standard or file/URL versions of PRISM variables.
}
\details{
You can see a full list of all variables using \code{data(airVars); airVars}.
}
\examples{
prVars('p')
prVars('s')
prVars('s', defs=TRUE)
}
\seealso{
airVars
}
