% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqMonths.r
\name{seqMonths}
\alias{seqMonths}
\title{Generate a sequence of years and months from a starting and ending date.}
\usage{
seqMonths(start, end)
}
\arguments{
\item{start, end}{Start and end dates. Objects of class \code{Date}, or of class \code{character} in either YYYY-MM-DD or YYYY-MO format (e.g., \code{2015-01-01} or \code{2015-01}; note the need for leading zeros for months <10).}
}
\value{
A character vector.
}
\description{
This function generates a sequence of years and months for a given start and end date. Days are dropped. For example, if the input is \code{2015-10-24} and \code{2016-02-26}, the output would be \code{2015-10}, \code{2015-11}, \code{2015-12}, \code{2016-01}, and \code{2015-02}.
}
\examples{
seqMonths('2015-01-01', '2017-02-26')
seqMonths('2017-02-26', '2015-01-01')
seqMonths('2015-10', '2017-02')
seqMonths('2015-10', '2015-02')
}
