% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stl.r
\name{stl}
\alias{stl}
\title{Outline of the City of Saint Louis, Missouri, USA}
\format{
\code{SpatVector}
}
\source{
\href{https://www.stlouis-mo.gov/}{City of Saint Louis}
}
\description{
This file is a \code{SpatVector} outline of the City of Saint Louis, Missouri, USA.
}
\examples{

stl <- airExample('stl')
stl
plot(stl)

}
\seealso{
\code{\link[=airExample]{airExample()}}
}
