% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcStack.r
\name{tcStack}
\alias{tcStack}
\title{Get a stack of TerraClimate rasters}
\usage{
tcStack(rastDir, vars, years)
}
\arguments{
\item{rastDir}{Name of the base path in which TerraClimate rasters are saved. It is assumed that inside the base path there are subfolders, one per TerraClimate variable, and inside these are NetCDF rasters representing "stacks" of monthly climate values, one per year. The assumed folder structure and file-naming scheme is exactly the same as that used by \code{\link[airUpThere]{tcDownload}}, so if the rasters were downloaded using that function, this function should work.}

\item{vars}{Name(s) of variable(s) to use.}

\item{years}{Year(s) of the period from which to apply the function.}
}
\value{
Object of class \code{SpatRaster}. Each raster in the stack will be named as per \if{html}{\out{<variable>}}\emph{\if{html}{\out{<year>}}}\if{html}{\out{<month>}}. For example, \code{tmin_1958_1} for the raster with minimum temperature for January of 1958. Standard variable names will be used (see \code{\link[airUpThere]{tcVars}}).
}
\description{
This function returns a stack of TerraClimate rasters for one or more variables for one or more years.
}
\examples{

\dontrun{
tcDir <- 'F:/ecology/Climate/TerraClimate/ORIGINALS'
tmins <- tcStack(tcDir, 'tmin', years=1971:1972)
tmins

}
}
