% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcVars.r
\name{tcVars}
\alias{tcVars}
\title{Get Names of TerraClimate variables}
\usage{
tcVars(standard = TRUE, defs = FALSE)
}
\arguments{
\item{standard}{If \code{TRUE} (default), the "standard" names are returned. If \code{FALSE}, the file/URL names are returned.}

\item{defs}{If \code{TRUE}, return general definitions of each variable, too.}
}
\value{
A character vector.
}
\description{
Get standard or file/URL versions of TerraClimate variables.
}
\details{
You can see a full list of all variables using \code{data(airVars); airVars}.
}
\examples{
tcVars(TRUE)
tcVars(TRUE, defs=TRUE)
tcVars(FALSE)
}
\seealso{
airVars
}
