% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.r
\name{trees}
\alias{trees}
\title{Trees in the City of Saint Louis, Missouri, USA}
\format{
\code{SpatVector}
}
\source{
\href{https://www.treesouis-mo.gov/}{City of Saint Louis}
}
\description{
This file is a \code{SpatVector} of a random sample of the locations of trees in Saint Louis, Missouri, USA.
}
\examples{

trees <- airExample('trees')
trees
plot(trees)

}
\seealso{
\code{\link[=airExample]{airExample()}}
}
