% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcInternalUtils.r
\name{wcCheckEsm_internal}
\alias{wcCheckEsm_internal}
\title{Check ESM for WorldClim}
\usage{
wcCheckEsm_internal(ver, esm)
}
\arguments{
\item{ver}{Version number (e.g., 1.4, 2.1)}

\item{Character, }{either short (2-letter) or long ESM names (case will not matter). Types cannot be mixed (i.e., all names must be 2-characters in length or long names).}
}
\value{
\code{TRUE} (invisible) or an error will be called.
}
\description{
Check ESM for WorldClim
}
\keyword{internal}
