% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcInternalUtils.r
\name{wcCheckGhg_internal}
\alias{wcCheckGhg_internal}
\title{Check if this scenario exists for WorldClim future version}
\usage{
wcCheckGhg_internal(ver, scenario)
}
\arguments{
\item{scenario}{Valid values for CMIP5 (RCPs) are one or more of 26, 45, 60, and/or 85.  Valid values for SSPs (CMIP6) are one or more of 126, 245, 370, and/or 585.}

\item{cmip}{5 or 6}
}
\value{
Logical
}
\description{
Check if this scenario exists for WorldClim future version
}
\keyword{internal}
