% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcInternalUtils.r
\name{wcCheckPeriod_internal}
\alias{wcCheckPeriod_internal}
\title{Check name of future time period for WorldClim}
\usage{
wcCheckPeriod_internal(ver, period, type = NULL)
}
\arguments{
\item{ver}{WorldClim version}

\item{period}{Year indicating time period (in "standard" format, like '2050' or '1960')}

\item{type}{Either of \code{future} or \code{decadal}. Only needed for WC 2.1.}
}
\value{
Character
}
\description{
Check name of future time period for WorldClim
}
\examples{
wcCheckPeriod_internal(1.4, 2050)
wcCheckPeriod_internal(1.4, 2070)
wcCheckPeriod_internal(2.1, 2050, 'future')
wcCheckPeriod_internal(2.1, 2111, 'future')
wcCheckPeriod_internal(2.1, 1960, 'decadal')
}
