% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcInternalUtils.r
\name{wcCheckVars_internal}
\alias{wcCheckVars_internal}
\title{Check to see if variables are valid WC variables}
\usage{
wcCheckVars_internal(ver, vars, period)
}
\arguments{
\item{ver}{WC version}

\item{vars}{Variables in standard format}

\item{period}{"historical", "future", "decadal" Partial matching OK.}
}
\value{
\code{TRUE} or fails
}
\description{
Check to see if variables are valid WC variables
}
\keyword{internal}
