% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcInternalUtils.r
\name{wcCheckVer_internal}
\alias{wcCheckVer_internal}
\title{Check WorldClim version number}
\usage{
wcCheckVer_internal(ver)
}
\arguments{
\item{ver}{WorldCli version number (e.g., 1.4, 2.1, etc.)}
}
\value{
Logical
}
\description{
Check WorldClim version number
}
\keyword{internal}
