% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcInternalUtils.r
\name{wcConvertEsm_internal}
\alias{wcConvertEsm_internal}
\title{Check and get ESM for WorldClim}
\usage{
wcConvertEsm_internal(ver, esm)
}
\arguments{
\item{ver}{1.4 or 2.1}

\item{esm}{Character}
}
\value{
Character
}
\description{
Check and get ESM for WorldClim
}
\keyword{internal}
