% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcConvertPeriod.r
\name{wcConvertPeriod}
\alias{wcConvertPeriod}
\title{Name of future time period in WorldClim URL and file names}
\usage{
wcConvertPeriod(ver, period, type = NULL, standardToFile = TRUE)
}
\arguments{
\item{ver}{WorldClim version: 1.4 or 2.1.}

\item{period}{Year indicating time period.}

\item{type}{Either of \code{future} or \code{decadal}. Only needed for WC 2.1.}

\item{standardToFile}{If \coed{TRUE} (default), then convert from "standard" to "file/URL" version. If \code{FALSE}, then from "file/URL" version to "standard."}
}
\value{
Character
}
\description{
Return the name of a future time WorldClim time period in the format used in URLs and file names.
}
\examples{
wcConvertPeriod(1.4, 2050)
wcConvertPeriod(1.4, 2070)
wcConvertPeriod(2.1, 2030)
wcConvertPeriod(2.1, 2050)
wcConvertPeriod(2.1, 2070)
wcConvertPeriod(2.1, 2090)
}
