% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcDownload.r
\name{wcDownloadHist}
\alias{wcDownloadHist}
\alias{wcDownloadFut}
\alias{wcDownloadDecadal}
\alias{wcDownloadElev}
\title{Download WorldClim climate rasters}
\usage{
wcDownloadHist(saveTo, ver, res, vars, overwrite = FALSE, verbose = TRUE)

wcDownloadFut(
  saveTo,
  ver,
  res,
  vars,
  esm,
  ghg,
  period,
  overwrite = FALSE,
  verbose = TRUE
)

wcDownloadDecadal(saveTo, vars, period, overwrite = FALSE, verbose = TRUE)

wcDownloadElev(saveTo, res, overwrite = FALSE, verbose = TRUE)
}
\arguments{
\item{saveTo}{Name of the base path to which to save the download. Subfolders will be created within this folder.}

\item{ver}{WorldClim Version. Either \code{1.4} or \code{2.1}. Note that some resolutions and/or combinations of variables and earth system models may not be available.}

\item{res}{Resolution of the rasters. Valid values include:
\itemize{
\item WC 1.4 and 2.1 historical normals: \code{10} (10 arcmin), 5\verb{ (5 arcmin), 2.5} (2.5 arcmin), and/or  \code{30} (30 arcsec).
\item WC 1.4 future scenario normals: \code{10} (10 arcmin), \code{5} (5 arcmin), \code{2.5} (2.5 arcmin), and/or \code{30} (30 arcsec).
\item WC 2.1 future scenario normals: \code{10} (10 arcmin), \code{5} (5 arcmin), and/or \code{2.5} (2.5 arcmin).
\item WC 2.1 decadal monthly averages: \code{2.5} (2.5 arcmin).
\item WC 2.1 elevation: \code{10} (10 arcmin), \code{5} (5 arcmin), \code{2.5} (2.5 arcmin), and/or  \code{30} (30 arcsec).
}}

\item{vars}{Name(s) of variable(s) to download. Different versions and time periods of WorldClim use different names for the same variable (e.g., "prec" versus "ppt" versus "pr" for precipitation). To reduce confusion, variable names have been standardized (in this package) to be the same across versions and times. Valid values are:
\itemize{
\item \code{tmin}: minimum temperature (available for WC 1.4 and 2.1 historical and future, and WC 2.1 decadal)
\item \code{tmax}: maximum temperature (available for WC 1.4 and 2.1 historical and future, and WC 2.1 decadal)
\item \code{tmean}: mean temperature (available for WC 1.4 and 2.1 historical)
\item \code{ppt}: precipitation (available for WC 1.4 and 2.1 historical and future, and WC 2.1 decadal)
\item \code{bio}: BIOCLIM variables (available for WC 1.4 and 2.1 historical and future)
\item \code{srad}: solar radiation (available for WC 2.1 historical)
\item \code{wind}: average wind speed (available for WC 2.1 historical)
\item \code{vap}: vapor pressure deficit (available for WC 2.1 historical)
\item \code{elev}: elevation (available for WC 2.1)
}}

\item{overwrite}{If \code{FALSE} (default), do not overwrite existing rasters.}

\item{verbose}{If \code{TRUE} (default), display progress.}

\item{esm}{Name(s) of one or more earth system models (global circulation models) for downloading future rasters. You can get the available names from \code{data(wcEsm)}. Valid values depend on whether WC version 1.4 or 2.1 is being called:
\itemize{
\item WorldClim 1.4 (CMIP5):
\itemize{
\item \verb{ACCESS1-0'} or \verb{AC'}
\item \verb{BCC-CSM1-1'} or \verb{BC'}
\item \verb{CCSM4'} or \verb{CC'}
\item \verb{CESM1-CAM5-1-FV2'} or \verb{CE'}
\item \verb{CNRM-CM5'} or \verb{CN'}
\item \verb{GFDL-CM3'} or \verb{GF'}
\item \verb{GFDL-ESM2G'} or \verb{GD'}
\item \verb{GISS-E2-R'} or \verb{GS'}
\item \verb{HadGEM2-AO'} or \verb{HD'}
\item \verb{HadGEM2-CC'} or \verb{HG'}
\item \verb{HadGEM2-ES'} or \verb{HE'}
\item \verb{INMCM4'} or \verb{IN'}
\item \verb{IPSL-CM5A-LR'} or \verb{IP'}
\item \verb{MIROC-ESM-CHEM'} or \verb{MI'}
\item \verb{MIROC-ESM'} or \verb{MR'}
\item \verb{MIROC5'} or \verb{MC'}
\item \verb{MPI-ESM-LR'} or \verb{MP'}
\item \verb{MRI-CGCM3'} or \verb{MG'}
\item \verb{NorESM1-M'} or \verb{NO'}
\item WorldClim 2.1 (CMIP6):
\item \verb{BCC-CSM2-MR'} or \verb{BC'}
\item \verb{CNRM-CM6-1'} or \verb{CC'}
\item \verb{CNRM-ESM2-1'} or \verb{CE'}
\item \verb{CanESM5'} or \verb{CA'}
\item \verb{GFDL-ESM4'} or \verb{GF'}
\item \verb{IPSL-CM6A-LR'} or \verb{IP'}
\item \verb{MIROC-ES2L'} or \verb{MR'}
\item \verb{MIROC6'} or \verb{MC'}
\item \verb{MRI-ESM2-0'} or \verb{ME'}
}
}}

\item{ghg}{Greenhouse gas emissions scenario for future rasters. Valid values depend on the version of WorldClim. One or more of a valid set can be specified. This argument is ignored if near present-day rasters are being downloaded.
* WorldClim 1.4 (CMIP5): These are representative concentration pathways (RCPs), and valid values are are one or more of 26, 45, 60, and/or 85.
\itemize{
\item WorldClim 2.1 (CMIP6): These are shared socioeconomic pathways (SSPs), and valid values are one or more of 126, 245, 370, and/or 585.
}}

\item{period}{Year(s) of the time period from which to download climate rasters. Valid values depend on the version of WorldClim and whether future or historical decadal averages are being downloaded:
\itemize{
\item WC 1.4 future scenario normals:
* \code{2050} (average across 2041-2060)
* \code{2070} (average across 2061-2080)
\item WC 2.1 future scenario normals:
* \code{2030} (average across 2021-2040)
* \code{2050} (average across 2041-2060)
* \code{2070} (average across 2061-2080)
* \code{2090} (average across 2081-2100)
\item WC 2.1 decadal monthly averages:
* \code{1960} (average across 1960-1969)
* \code{1970} (average across 1970-1979)
* \code{1980} (average across 1980-1989)
* \code{1990} (average across 1990-1999)
* \code{2000} (average across 2000-2000)
* \code{2010} (average across 2010-2018)
}}
}
\value{
One or more zipped raster sets are saved to disk. The function also returns a data frame indicating if the desired file(s) were already on the disk and if they were downloaded.
}
\description{
These functions download WorldClim climate rasters:
\itemize{
\item \code{wcDownloadHist}: Historical (near-present) climate normals from WorldClim versions 1.4 and 2.1.
\item \code{wcDownloadFut}: Future climate scenario rasters from WorldClim versions 1.4 and 2.1.
\item \code{wdDownloadDecade}: Monthly averages for particular historical decades from WorldClim version 2.1.
\item \code{ecDownloadElev}: Elevation from WorldClim version 2.1.
}
}
\section{Functions}{
\itemize{
\item \code{wcDownloadFut()}: Download WorldClim future normal rasters

\item \code{wcDownloadDecadal()}: Download WorldClim historical decadal average rasters

\item \code{wcDownloadElev()}: Download WorldClim elevation rasters

}}
\examples{

\dontrun{
dl1_4 <- 'C:/ecology/!Scratch/wc 1.4 zip'
dl2_1 <- 'C:/ecology/!Scratch/wc 2.1 zip'

# historical (near-present)
wcDownloadHist(dl1_4, ver=1.4, res=10, vars=c('tmin', 'tmax'))
wcDownloadHist(dl2_1, ver=2.1, res=10, vars=c('tmin', 'tmax'))

# future: using BCC-CSM1-1 and BCC-CSM2-MR ESMs for RCP 6.0 and SSP 370
wcDownloadFut(dl1_4, ver=1.4, res=10, vars='tmin', esm='BC', ghg=60, period=2050)
wcDownloadFut(dl2_1, ver=2.1, res=10, vars='tmin', esm='BC', ghg=370, period=2050)

# historical decadal: for 1960-1969
wcDownloadDecadal(dl2_1, vars='tmin', period=1960)

# elevation
wcDownloadElev(dl2_1, res=10)

}
}
\references{
Fick, S.E. and Hijmans, R.J. 2017. WorldClim 2: New 1-km spatial resolution climate surfaces for global land areas. \emph{International Journal of Climatology} 37:4302-4315. \doi{10.1002/joc.5086}

Hijmans, R.J., Cameron, S.E., Parra, J.L., Jones, P.G., and Jarvis, A. 2005. Very high resolution interpolated climate surfaces for global land areas. \emph{International Journal of Climatology} 25:1965-1978. \doi{10.1002/joc.1276}.
}
