% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcRes.r
\docType{data}
\name{wcRes}
\alias{wcRes}
\title{Names resolutions available for WorldClim}
\format{
An object of class \code{'data.frame'}.
}
\usage{
data(wcRes)
}
\description{
Names of WorldClim resolutions in "standard" and "file/URL" formats.
}
\examples{
data(wcRes)
wcRes
}
\seealso{
\code{\link[airUpThere]{wcConvertRes}}
}
\keyword{variables}
