% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcVars.r
\name{wcVars}
\alias{wcVars}
\title{Get Names of WorldClim variables}
\usage{
wcVars(ver, period, standard = TRUE, defs = FALSE)
}
\arguments{
\item{ver}{WorldClim version number: 1.4 or 2.1.}

\item{period}{Any of \code{'historical'}, \code{'future'}, or \code{'decadal'}  Partial matching is supported.}

\item{standard}{If \code{TRUE} (default), the "standard" names are returned. If \code{FALSE}, the file/URL names are returned.}

\item{defs}{If \code{TRUE}, return general definitions of each variable, too.}
}
\value{
A character vector.
}
\description{
Get standard or file/URL versions of WorldClim variables.
}
\details{
You can see a full list of all variables using \code{data(airVars); airVars}.
}
\examples{
wcVars(1.4, 'historical')
wcVars(1.4, 'historical', FALSE)
wcVars(1.4, 'historical', defs=TRUE)

wcVars(1.4, 'future')
wcVars(1.4, 'future', FALSE)

wcVars(2.1, 'historical')
wcVars(2.1, 'historical', FALSE)
wcVars(2.1, 'future')
wcVars(2.1, 'future', FALSE)

wcVars(2.1, 'decadal')

}
\seealso{
airVars
}
