% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgs84.r
\name{wgs84}
\alias{wgs84}
\title{Well-Known Text string for the WGS84 Datum}
\format{
\code{character}
}
\description{
This is the Well-Known Text (WKT) string for the World Geodetic System 1984 (WGS84).
}
\examples{

data(wgs84)
wgs84
cat(wgs84)

}
\seealso{
\link{nad83}
}
