% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prExtractRelative.r
\name{prismExtractRelative}
\alias{prismExtractRelative}
\alias{prExtractRelativeDaily}
\alias{prExtractRelativeMonthly}
\alias{prExtractRelativeAnnual}
\title{Extract values from PRISM across a time period relative to specific dates}
\usage{
prExtractRelativeDaily(
  prDir,
  x,
  vars,
  date,
  res = 4,
  rastSuffix = "bil",
  longLat = NULL,
  windowYears = 0,
  windowDays = 0,
  verbose = TRUE,
  ...
)

prExtractRelativeMonthly(
  prDir,
  x,
  vars,
  date,
  res = 4,
  rastSuffix = "bil",
  longLat = NULL,
  windowYears = 0,
  windowMonths = 0,
  verbose = TRUE,
  ...
)

prExtractRelativeAnnual(
  prDir,
  x,
  vars,
  date,
  res = 4,
  rastSuffix = "bil",
  annualDir = "annual",
  longLat = NULL,
  windowYears = 0,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{prDir}{Character. Path of the base directory in which the rasters are stored. It must have a structure with subfolders as described above.}

\item{x}{"Points" object of class \code{SpatVector}, \code{SpatialPoints}, \code{SpatialPointsDataFrame}, \code{data.frame}, or \code{matrix}.}

\item{vars}{Name of variable(s) to extract. Valid values are:
\itemize{
    \item    \code{ppt}: Precipitation.
    \item    \code{tmax}: Maximum temperature.
    \item    \code{tmin}: Minimum temperature.
    \item    \code{tmean}: Mean temperature.
    \item    \code{tdmean}: Dew-point temperature.
    \item    \code{vpdmax}: Maximum vapor-pressure.
    \item    \code{vpdmin}: Minimum vapor-pressure.
}}

\item{date}{Either:
\itemize{
    \item     Object of class \code{Date}. You can name a single date, in which case this date will be used for all records, or a vector of dates, one per point in \code{x}.
    \item     Name of column in \code{x} with date of each record. Values must be of in YYYY-MM-DD (year-month-day of month) format \emph{or} already of class \code{\link{Date}}. See \code{\link[lubridate]{ymd}} or related functions for more help.
    \item   Names of columns in \code{x} with year, month, and day of month of each record (in that order). Months must be numeric (i.e., 10, not "October").
}}

\item{res}{Resolution of the rasters. Valid values are either \code{30} or \code{800} (these are treated as the same: 30-arcsecond also known as the "800-m" resolution version of PRISM), \emph{or} code{1} or \code{4} (these are also treated as the same: the 1-arcminute also known as the "4-km" resolution version of PRISM).}

\item{rastSuffix}{Character. The "suffix" at the end of each weather raster. PRISM rasters are usually shipped in 'BIL' format, so normally this should be \code{bil}. However, any other suffix corresponding to a raster type that can be opened by the \code{\link[terra]{rast}} function can be used if the rasters have been converted to another format.}

\item{longLat}{Character vector with two elements. If \code{x} is an object of class \code{data.frame}, this is the name of the columns in \code{x} with longitude and latitude (in that order). Coordinates will be assumed to be in the same coordinate reference system as the PRISM rasters. This argument is ignored if \code{x} is not a data frame.}

\item{windowYears, windowMonths, windowDays}{Integers >= 0. Extract data for this many years, months, and/or days before the day of each observation, \emph{plus} the day of observation. Note:
\itemize{
    \item For daily data, only \code{windowYears} and \code{windowDays} are used. Note that the difference between using, say, \code{windowYears = 1} and \code{windowDays = 365} is that the former can accommodate leap days whereas the latter just extracts the 365 days prior (which may be less than a full calendar year if the timespan encompasses a leap day).
    \item For monthly data, only \code{windowYears} and \code{windowMonths} are used.
    \item For annual data, only \code{windowYears} is used.
}
To get only data for the day/month/year of each record, set all of the respective \code{window} arguments to 0 (default).}

\item{verbose}{Logical. If \code{TRUE} (default), show progress.}

\item{...}{Argument(s) to pass to \code{\link[terra]{extract}}.}

\item{annualDir}{Name of the highest-level folder in which annual rasters are stored. If you use the \code{prDownloadAnnual} function to download PRISM rasters, this will be \code{'annual'} (default). However, if you are extracting values from a purchased version of PRISM (i.e., that the PRISM staff sent you on a hard drive), the annual rasters are typically stored in the folder called \code{'monthly'}.}
}
\value{
Matrix with one row per row in \code{x}. \code{NA} values represent days/months/years that did not fall within the specified window or for which rasters were unavailable.
}
\description{
These functions extract values from interpolated weather rasters from the Parameter Regression on Independent Slopes (PRISM) data product across a window of time relative to a given date. For example, it could extract all values starting 10 days prior to a given date, with the date varying by point. If you wish to extract values across a specified date range (e.g., from 2014-04-22 to 2014-04-29, then see \code{\link{prExtractAbsolute}}.  Extractions are done at points (versus polygons or lines, for example). \cr\cr
The basic input is an object of class \code{spatVector}, \code{SpatialPoints}, \code{SpatialPointsDataFrame}, \code{data.frame}, or \code{matrix}, with each row representing a point location. The function also needs to be pointed toward a folder with PRISM data. The folder with PRISM data must be structured as:
\itemize{
	\item Base folder (e.g., \code{'C:/PRISM/an81'} or \code{'C:/PRISM/lt81'}), which contains:
	\item A folder named either \code{daily}, \code{monthly}, \code{annual}, which contains:
	\item One folder per variable (e.g., \code{tmin}, \code{tmax}, \code{vpdmax}), each of which contain:
	\item One folder per year (e.g., \code{1981}, code{1982}, etc.), each of which rasters like:
	\itemize{
		\item One raster per day named like \code{prism_tmin_us_30s_19810101.bil}, \code{prism_tdmean_us_30s_19810102.bil}, etc.
		\item One raster per month named like \code{prism_tmin_us_30s_198101.bil}, \code{prism_tdmean_us_30s_198101.bil}, etc.
		\item One raster representing the annual value named like \code{prism_tmin_us_30s_1981.bil}.
		}
}
\cr
The function can extract values corresponding to the day/month/year of each record, plus (optionally) a user-specified window of time prior to the day/month/year of each record. For example, you could use this to extract daily climate data for a site collected on April 22, 2014, and all days prior up to 10 years (April 23, 2004). This function is really a fancy wrapper for \code{\link[terra]{extract}}, but it does the tough job of getting the directory structures right, pulling all needed rasters, and efficiently grouping records to speed extraction. \cr\cr
The function does not assume that data for all PRISM years are available, but it does assume that all relevant rasters for a particular year are available within each yearly folder. If rasters preceding a date only partially cover the window, then values for the part covered will be extracted. For example if you try to extract annual values for a window spanning 2010 to 2020 but only have available rasters for 1981 to 2018, then values for 2010 to 2018 will be extracted. Values that cannot be extracted are represented by \code{NA} in the output.
}
\section{Functions}{
\itemize{
\item \code{prExtractRelativeDaily}: Extract daily values from PRISM across a temporal window

\item \code{prExtractRelativeMonthly}: Extract monthly values from PRISM across a temporal window

\item \code{prExtractRelativeAnnual}: Extract annual values from PRISM across a temporal window
}}

\examples{
\dontrun{
 x <- data.frame(
 	long=rep(-97.66, 5),
 	lat=rep(38.37, 5),
 	date=c('2015-12-31', '1981-01-05', '2020-12-01',
 	'2019-01-05', '1895-05-01')
)
 
 
# NA values are unavailable
prDir <- 'F:/ecology/Climate/PRISM/acquired_2020/an81'
y <- prExtractRelativeDaily(
 	x,
 	date = 'date',
 	longLat = c('long', 'lat'),
 	prDir = prDir,
 	vars = 'tmin',
		res = 30,
 	rastSuffix = 'tif',
 	windowYears = 0,
 	windowDays = 7,
 	verbose = TRUE
)
 
# NA values are unavailable
prDir <- 'F:/ecology/Climate/PRISM/acquired_2020/lt81'
y <- prExtractRelativeMonthly(
 	x,
 	date = 'date',
 	longLat = c('long', 'lat'),
 	prDir = prDir,
 	vars = 'tmin',
		res = 30,
 	rastSuffix = 'tif',
 	windowYears = 0,
 	windowMonths = 5,
 	verbose = TRUE
)
 
# NA values are unavailable
prDir <- 'F:/ecology/Climate/PRISM/acquired_2020/lt81'
y <- prExtractRelativeAnnual(
 	x,
 	date = 'date',
		longLat = c('long', 'lat'),
		prDir = prDir,
		vars = 'tmin',
		res = 30,
		rastSuffix = 'tif',
		annualDir = 'monthly',
		windowYears = 3,
		verbose = TRUE
)

}
}
