% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prStack.r
\name{prStack}
\alias{prStack}
\title{Create a raster stack from PRISM rasters}
\usage{
prStack(
  prDir,
  vars,
  dates,
  by = "day",
  span = FALSE,
  res = 4,
  rastSuffix = "bil",
  annualDir = "annual"
)
}
\arguments{
\item{prDir}{Character. Path of the base directory in which the rasters are stored.}

\item{vars}{Name of variable(s) to stack. Valid values are:
\itemize{
\item \code{ppt}: Precipitation.
\item \code{tmax}: Maximum temperature.
\item \code{tmin}: Minimum temperature.
\item \code{tmean}: Mean temperature.
\item \code{tdmean}: Dew-point temperature.
\item \code{vpdmax}: Maximum vapor-pressure.
\item \code{vpdmin}: Minimum vapor-pressure.
}}

\item{span}{If \code{FALSE} (default), then use the exact dates in \code{dates}. If \code{TRUE}, then extract a time series.}

\item{res}{Resolution of the rasters. Valid values are either \code{30} or \code{800} (i.e., 30-arcsecond or the "800-m" resolution version of PRISM), \emph{or} \code{1} or \code{4} (i.e., the 1-arcminute or "4-km" resolution version of PRISM).}

\item{rastSuffix}{Character. The "suffix" at the end of each weather raster. PRISM rasters are usually shipped in 'BIL' format, so normally this should be \code{bil}. However, any other suffix corresponding to a raster type that can be opened by the \code{\link[terra:rast]{terra::rast()}} function can be used if the rasters have been converted to another format.}

\item{annualDir}{Name of the highest-level folder in which annual rasters are stored. If you use the \code{prDownloadAnnual()} function to download PRISM rasters, this will be \code{'annual'} (default). However, if you are extracting values from a purchased version of PRISM (i.e., that the PRISM staff sent you on a hard drive), the annual rasters are typically stored in the folder called \code{'monthly'}. This is not needed when stacking daily or monthly rasters.}
}
\value{
Object of class \code{rast}.
}
\description{
This function creates a raster stack from select PRISM rasters. The stack consists of one or more variables across one or more time periods.
}
\examples{
\dontrun{
prDir <- 'F:/ecology/Climate/PRISM/acquired_2020/an81'

# specific days
y <- prStack(
 	prDir = prDir,
 	vars = c('tmin', 'tmax'),
		dates = c('2015-01-01', '2015-01-05'),
		by = 'day',
		res = 30,
 	rastSuffix = 'tif'
)

# series of days
y <- prStack(
 	prDir = prDir,
 	vars = c('tmin', 'tmax'),
		dates = c('2015-01-01', '2015-01-05'),
		by = 'day',
		span = TRUE,
		res = 30,
 	rastSuffix = 'tif'
)

prDir <- 'F:/ecology/Climate/PRISM/acquired_2020/lt81'
# specific months
y <- prStack(
 	prDir = prDir,
 	vars = c('tmin', 'tmax'),
		dates = c('2015-01-01', '2015-05-31'),
		by = 'month',
		res = 30,
 	rastSuffix = 'tif'
)

# series of months
y <- prStack(
 	prDir = prDir,
 	vars = c('tmin', 'tmax'),
		dates = c('2015-01-01', '2015-05-31'),
		by = 'month',
		span = TRUE,
		res = 30,
 	rastSuffix = 'tif'
)
 
# specific years
y <- prStack(
 	prDir = prDir,
 	vars = c('tmin', 'tmax'),
		dates = c('2015-01-01', '2018-05-31'),
		by = 'year',
 	rastSuffix = 'tif',
		res = 30,
		annualDir = 'monthly'
)

# series of months
y <- prStack(
 	prDir = prDir,
 	vars = c('tmin', 'tmax'),
		dates = c('2015-01-01', '2018-05-31'),
		by = 'year',
		span = TRUE,
 	rastSuffix = 'tif',
		res = 30,
		annualDir = 'monthly'
)

}
}
