% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcDownload.r
\name{tcDownload}
\alias{tcDownload}
\alias{tcDownloadMonthly}
\alias{tcDownloadElev}
\title{Download TerraClimate climate rasters}
\usage{
tcDownloadMonthly(
  saveTo,
  vars,
  year = NULL,
  elevation = FALSE,
  update = TRUE,
  forceUpdate = FALSE,
  verbose = TRUE
)

tcDownloadElev(saveTo, update = TRUE, forceUpdate = FALSE, verbose = TRUE)
}
\arguments{
\item{saveTo}{Name of the base path to which to save the download. Subfolders will be created within this folder.}

\item{vars}{Name(s) of variable(s) to download. Valid values include:

\if{html}{\out{<div class="sourceCode">}}\preformatted{* `elev`: elevation
* `tmin`: minimum temperature
* `tmax`: maximum temperature
* `ppt`: accumulated precipitation
* `swe`: snow water equivalent
* `aet`: actual evapotranspiration
* `pet`: reference (potential) evapotranspiration (ET0)
* `def`: climate water deficit
* `pdsi`: Palmer Drought Severity Index
* `soilWater`: extractable soil moisture
* `srad`: downward shortwave flux at the surface (solar radiation)
* `windSpeed`: average wind speed
* `q`: cumulative streamflow
* `vap`: vapor pressure
* `vpd`: vapor pressure deficit
}\if{html}{\out{</div>}}

`}

\item{year}{Year(s) of the period from which to download monthly climate rasters. Valid years include 1958 through the present (often one or two years prior to the current year). This can be \code{NULL} only if the variable being downloaded is \code{elevation}.}

\item{elevation}{Logical: If \code{TRUE}, download elevation, too.}

\item{update}{If \code{TRUE} (default), only rasters that have been generated after the one(s) on disk will be downloaded. The rasters being updated will be overwritten. If \code{FALSE}, only rasters that are missing will be downloaded (existing rasters will not be overwritten).}

\item{forceUpdate}{If \code{FALSE} (default), then all existing rasters will be overwritten, regardless of whether or not the existing ones are up-to-date or not.}

\item{verbose}{If \code{TRUE} (default), display progress.}
}
\value{
NetCDF rasters are saved to disk. The function also returns a data frame indicating if the desired file(s) were already on the disk and if they were downloaded.
}
\description{
These functions download \href{http://www.climatologylab.org/terraclimate.html}{TerraClimate} climate rasters and elevation.
}
\section{Functions}{
\itemize{
\item \code{tcDownloadElev()}: Download TerraClimate elevation raster

}}
\examples{

\dontrun{
dl <- 'C:/ecology/!Scratch/tc'
tcDownloadMonthly(dl, 'tmin', 1958)
tcDownloadElev(dl, 'elev')

}
}
\references{
Abatzoglou, J.T., Dobrowski, S.Z., Parks, S.A., and Hegewisch, K.C. 2018. TerraClimate, a high-resolution global dataset of monthly climate and climatic water balance from 1958-2015. \emph{Scientific Data} 5:170191. \doi{10.1038/sdata.2017.191}.
}
