% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcConvertGhg.r
\name{wcConvertGhg}
\alias{wcConvertGhg}
\title{Nice name for GHG WorldClim scenarios}
\usage{
wcConvertGhg(ver, ghg)
}
\arguments{
\item{ver}{Version number: 1.4 or 2.1}

\item{ghg}{Greenhouse gas emissions scenario as a \coed{numeric} value. This is the RCP or SSP number (without a decimal point, for RCPs).}
}
\value{
Character
}
\description{
Create a nice name for a WorldClim future climate scenario.
}
\examples{
wcConvertGhg(1.4, 4.5)
wcConvertGhg(1.4, 8.5)
wcConvertGhg(2.1, 126)
wcConvertGhg(2.1, 370)
}
