% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcConvertRes.r
\name{wcConvertRes}
\alias{wcConvertRes}
\title{Convert name resolution of WorldClim in "standard" and URL/file formats}
\usage{
wcConvertRes(ver, res, period, standardToFile = TRUE, incUnit = TRUE)
}
\arguments{
\item{ver}{Either \code{1.4} or \coed{2.1}.}

\item{res}{Either "nice" names of resolutions (any of \code{10}, \code{5}, \code{2.5}, or \code{30}), or "file/URL" names:
\itemize{
\item WorldClim Version 1.4, historical: \code{10m}, \code{5m}, \code{2-5m}, or \code{30s}
\item WorldClim Version 1.3, future: \code{10m}, \code{5m}, \code{2_5m}, or \code{30s}
\item WorldClim Version 2.1: \code{10m}, \code{5m}, \code{2.5m}, or \code{30s}
}}

\item{period}{Either \code{'historical'}, \code{'future'}, or \code{'decadal'}.}

\item{standardToFile}{If \code{TRUE}, \code{res} is assumed to be in "standard" format, so will be translated into "file/URL" format. If \code{FALSE}, then it will be assumed to be in "file/URL" format and will be translated into "standard" format (possibly with unit, if \code{incUnit} is \code{TRUE}).}

\item{incUnit}{If \code{TRUE}, include unit if \code{standardToFile} is also \code{TRUE}. If \code{standardToFile} is \code{FALSE}, this is ignored.}
}
\value{
Character
}
\description{
Convert between the "standard" and "file/URL" versions of names of resolutions in WorldClim. You can see a full table of resolutions using \code{data(wcRes)}.
}
\examples{
wcConvertRes(1.4, 2.5, 'historical', TRUE)
wcConvertRes(1.4, '2-5m', 'historical', FALSE)
wcConvertRes(1.4, 2.5, 'future', TRUE)
wcConvertRes(1.4, '2_5m', 'future', FALSE)
wcConvertRes(2.1, 2.5, 'historical', TRUE)
wcConvertRes(2.1, '2.5m', 'historical', FALSE)
}
